requirements.txt
flask==2.3.3
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.1
openpyxl==3.1.2
